Fixing bugs, implementing new BM25 algorithm variant, changing score file offsets to longs.

+++ src/main/java/com/dreamcloud/esa_score/fs/DocumentScoreDataReader.java
@@ -4,5 +4,5 @@ import java.io.IOException;
 import java.nio.ByteBuffer;
 
 public interface DocumentScoreDataReader {
-    ByteBuffer readScores(int offset, int numScores) throws IOException;
+    ByteBuffer readScores(long offset, int numScores) throws IOException;
 }
